Move react-toastify code and onProgress and onLoad callbacks from @stlite/sharing to @stlite/common-react
