+++ packages/common-react/src/toastify-components/ToastContainer.tsx
@@ -1,0 +1,13 @@
+import { ToastContainer as DefaultToastContainer } from "react-toastify";
+import { isDarkTheme } from "./theme";
+
+function ToastContainer() {
+  return (
+    <DefaultToastContainer
+      style={{ zIndex: 999999 }}
+      theme={isDarkTheme() ? "dark" : "light"}
+    />
+  );
+}
+
+export default ToastContainer;
