+++ packages/sharing/src/App.tsx
@@ -7,9 +7,10 @@ import {
   ReplyMessage,
 } from "@stlite/sharing-common";
 import StreamlitApp from "./StreamlitApp";
-import { toast, Slide, Id as ToastId } from "react-toastify";
+import { toast } from "react-toastify";
 import ErrorToastContent from "./components/ErrorToastContent";
 import "./App.css";
+import { makeToastKernelCallbacks } from "@stlite/common-react";
 
 const editorAppOriginRegex = process.env.REACT_APP_EDITOR_APP_ORIGIN_REGEX
   ? new RegExp(process.env.REACT_APP_EDITOR_APP_ORIGIN_REGEX)
@@ -71,29 +72,7 @@ st.write("Hello World")`,
 
         console.debug("Initialize with", appData);
 
-        let prevToastId: ToastId | null = null;
-        const toastIds: ToastId[] = [];
-        const onProgress: StliteKernelOptions["onProgress"] = (message) => {
-          const id = toast(message, {
-            position: toast.POSITION.BOTTOM_RIGHT,
-            transition: Slide,
-            isLoading: true,
-            hideProgressBar: true,
-            closeButton: false,
-          });
-          toastIds.push(id);
-
-          if (prevToastId) {
-            toast.update(prevToastId, {
-              isLoading: false,
-              autoClose: 3000,
-            });
-          }
-          prevToastId = id;
-        };
-        const onLoad: StliteKernelOptions["onLoad"] = () => {
-          toastIds.forEach((id) => toast.dismiss(id));
-        };
+        const { onProgress, onLoad } = makeToastKernelCallbacks();
         const onError: StliteKernelOptions["onError"] = (error) => {
           toast(
             <ErrorToastContent
