+++ packages/sharing/src/index.tsx
@@ -2,17 +2,13 @@ import React from "react";
 import ReactDOM from "react-dom";
 import App from "./App";
 import reportWebVitals from "./reportWebVitals";
-import { ToastContainer } from "react-toastify";
+import { ToastContainer } from "@stlite/common-react";
 import "react-toastify/dist/ReactToastify.css";
-import { isDarkTheme } from "./theme";
 
 ReactDOM.render(
   <React.StrictMode>
     <App />
-    <ToastContainer
-      style={{ zIndex: 999999 }}
-      theme={isDarkTheme() ? "dark" : "light"}
-    />
+    <ToastContainer />
   </React.StrictMode>,
   document.getElementById("root") as HTMLElement
 );
